+++ flatlaf-testing/src/main/java/com/formdev/flatlaf/testing/FlatNativeWindowBorderTest.java
@@ -35,8 +35,8 @@ import com.formdev.flatlaf.FlatIntelliJLaf;
 import com.formdev.flatlaf.FlatLaf;
 import com.formdev.flatlaf.FlatLightLaf;
 import com.formdev.flatlaf.extras.FlatInspector;
+import com.formdev.flatlaf.extras.components.FlatTriStateCheckBox;
 import com.formdev.flatlaf.ui.FlatLineBorder;
-import com.formdev.flatlaf.ui.FlatNativeWindowBorder;
 import com.formdev.flatlaf.util.SystemInfo;
 import net.miginfocom.swing.*;
 
@@ -304,13 +304,11 @@ public class FlatNativeWindowBorderTest
 	}
 
 	private void nativeChanged() {
-		FlatNativeWindowBorder.setHasCustomDecoration( window, nativeCheckBox.isSelected() );
+		FlatLaf.setUseNativeWindowDecorations( nativeCheckBox.isSelected() );
 	}
 
 	private void native2Changed() {
-		((RootPaneContainer)window).getRootPane().putClientProperty( FlatClientProperties.USE_WINDOW_DECORATIONS, native2CheckBox.isSelected() );
-		window.dispose();
-		window.setVisible( true );
+		((RootPaneContainer)window).getRootPane().putClientProperty( FlatClientProperties.USE_WINDOW_DECORATIONS, native2CheckBox.getChecked() );
 	}
 
 	private void revalidateLayout() {
@@ -382,7 +380,7 @@ public class FlatNativeWindowBorderTest
 		undecoratedCheckBox = new JCheckBox();
 		fullScreenCheckBox = new JCheckBox();
 		nativeCheckBox = new JCheckBox();
-		native2CheckBox = new JCheckBox();
+		native2CheckBox = new FlatTriStateCheckBox();
 		openDialogButton = new JButton();
 		hideWindowButton = new JButton();
 		reopenButton = new JButton();
@@ -446,7 +444,6 @@ public class FlatNativeWindowBorderTest
 
 		//---- native2CheckBox ----
 		native2CheckBox.setText("JRootPane.useWindowDecorations");
-		native2CheckBox.setSelected(true);
 		native2CheckBox.addActionListener(e -> native2Changed());
 		add(native2CheckBox, "cell 0 3 3 1");
 
@@ -507,7 +504,7 @@ public class FlatNativeWindowBorderTest
 	private JCheckBox undecoratedCheckBox;
 	private JCheckBox fullScreenCheckBox;
 	private JCheckBox nativeCheckBox;
-	private JCheckBox native2CheckBox;
+	private FlatTriStateCheckBox native2CheckBox;
 	private JButton openDialogButton;
 	private JButton hideWindowButton;
 	private JButton reopenButton;